Move update logic to compliance handler component
